% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_conversion.R
\name{cGPD2Omori}
\alias{cGPD2Omori}
\title{Convert ETAS parameter vector from cGPD to Omori-law parameterisation.}
\usage{
cGPD2Omori(cGPDPars, Mbar, M0)
}
\arguments{
\item{cGPDPars}{Vector of Omori-law ETAS parameters (mu, C, alpha, nu_t, xi_t)}

\item{Mbar}{Mean magnitude of catalogue}

\item{M0}{Threshold magnitude of catalogue}
}
\value{
Vector of cGPD parameters (mu, K, alpha, c, p)
}
\description{
Convert ETAS parameter vector from cGPD to Omori-law parameterisation.
}
