% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrETAS-package.R
\docType{package}
\name{corrETAS}
\alias{corrETAS}
\title{Bayesian Estimation of the ETAS Model for Earthquakes with Dependent Magnitudes}
\usage{
corrETAS()
}
\description{
Bayesian Estimation of the ETAS Model for Earthquakes with Dependent Magnitudes
}
\references{
Gordon J. Ross - Bayesian Estimation of the ETAS Model for Earthquake Occurrences (2016)
}
\author{
Zak Varty \email{zdvarty@gmail.com}
}
