% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_gpd_to_gauss.R
\name{cpp_gpd_to_gauss}
\alias{cpp_gpd_to_gauss}
\title{Conversion from GPD to Gaussian margin in C++}
\usage{
cpp_gpd_to_gauss(x_gpd, nu, xi, mu)
}
\arguments{
\item{x_gpd}{Numeric vector of gpd observations}

\item{nu}{gpd alt-scale parameter}

\item{xi}{gpd shape parameter}

\item{mu}{gpd threshold}
}
\value{
NULL prints result of Cpp function to put x_gpd on gaussian margins
}
\description{
Conversion from GPD to Gaussian margin in C++
}
