% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_gpd_to_gauss.R
\name{cpp_pgpd}
\alias{cpp_pgpd}
\title{Testing C++ GPD cdf calculation}
\usage{
cpp_pgpd(q, nu, xi, mu)
}
\arguments{
\item{q}{Numeric vector of values at which to print CDF}

\item{nu}{gpd alt-scale parameter}

\item{xi}{gpd shape parameter}

\item{mu}{gpd threshold}
}
\value{
NULL prints CDF value(s) at q.
}
\description{
Testing C++ GPD cdf calculation
}
