% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_rdiscrete.R
\name{cpp_rdiscrete}
\alias{cpp_rdiscrete}
\title{Sample from a discrete distribution in C++}
\usage{
cpp_rdiscrete(probs, n)
}
\arguments{
\item{probs}{vector of non-negative sampling weights}

\item{n}{number of samples to draw from distribtuion}
}
\value{
n samples from discrete distribution
}
\description{
Sample from a discrete distribution in C++
}
\examples{
cpp_rdiscrete(probs = c(1,2.5,5), n = 20)
}
