% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_pareto_distribution.R
\name{dgpd}
\alias{dgpd}
\title{Generalised Pareto Distribution}
\usage{
dgpd(x, shape, scale = NULL, nu = NULL, mu = 0, log = FALSE)
}
\arguments{
\item{x}{vector of values as which to evaluate density.}

\item{shape}{shape parameter (xi)}

\item{scale}{scale parameter (sigma)}

\item{nu}{alternative scale parameter}

\item{mu}{location parameter}

\item{log}{locical. Return log}
}
\value{
density of the GPD at x
}
\description{
Density function of the GPD specified in terms of (sigma,xi) or (nu,xi).
Improvement over evir function as it returns an error if the (implied) shape
parameter is non-positive. Also properly handles cases where and xi=0 or x is
outside of the domain of the given distribution.
}
\examples{
dgpd(x = c(-1,0.5,1,1.9,5),shape = -0.5, scale = 1)
}
\author{
Zak Varty
}
