% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateETAScorr.R
\name{draw_M_after}
\alias{draw_M_after}
\title{Draw aftershock magnitudes from a distribution that is dependent on the mainshock magnitude.}
\usage{
draw_M_after(
  M_main,
  n_after,
  nuxi_main,
  nuxi_after,
  rho = 0,
  mu_main = 0,
  mu_after = 0
)
}
\arguments{
\item{M_main}{Magnitude of main shock}

\item{n_after}{Number of aftershock magnitudes to draw}

\item{nuxi_main}{Parameters of the marginal mainshock GPD.}

\item{nuxi_after}{Parameters of the marginal aftershock GPD.}

\item{rho}{Correlation coefficient of gaussian copula.}

\item{mu_main}{GPD threshold for mainshocks, must be smaller than M_main.}

\item{mu_after}{GPD threshold for aftershocks.}
}
\value{
A vector of sampled aftershock magnitudes.
}
\description{
Draw aftershock magnitudes from a distribution that is dependent on the mainshock magnitude.
}
\examples{
nuxi_main = c(2,0.2)
nuxi_after = c(0.5,0)
draw_M_after(M_main = 2.12, n_after = 30, nuxi_main = nuxi_main, nuxi_after = nuxi_after, rho =0.7)
}
