% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAS_intensity.R
\name{ETAS_intensity}
\alias{ETAS_intensity}
\title{Calculate the conditional intenisty function of an ETAS point process}
\usage{
ETAS_intensity(t, ts, ms, cGPDPars = NULL, OmoriPars = NULL, M0 = NULL)
}
\arguments{
\item{t}{Vector of times at which to calculate the intensity.}

\item{ts}{Vector of observed event times.}

\item{ms}{Vector of observed event magnitudes.}

\item{cGPDPars}{Vector of ETAS parameters (mu,C,a,nu_t,xi_t).}

\item{OmoriPars}{Alternative vector of ETAS parameters (mu,K,a,c,p).}

\item{M0}{Minimum triggering magnitude.}
}
\value{
Vector of conditional intensities.
}
\description{
Calculate the conditional intenisty function of an ETAS point process
}
