% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_pareto_likelihoods.R
\name{exactly_one_null}
\alias{exactly_one_null}
\title{Check that exactly one of two objects is NULL.}
\usage{
exactly_one_null(a, b)
}
\arguments{
\item{a}{first object}

\item{b}{second object}
}
\value{
logical. Is exactly one a and b NULL?
}
\description{
Check that exactly one of two objects is NULL.
}
\examples{
exactly_one_null(3, NULL)
}
