% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_pareto_likelihoods.R
\name{gpd_flat_prior}
\alias{gpd_flat_prior}
\title{Evaluate flat prior density for gpd parameters}
\usage{
gpd_flat_prior(
  sig = NULL,
  nu = NULL,
  xi,
  mu = 0,
  logged = FALSE,
  optim_friendly = FALSE
)
}
\arguments{
\item{sig}{gpd scale parameter}

\item{nu}{gpd alt-scale parameter}

\item{xi}{gpd shape parameter}

\item{mu}{gpd threshold}

\item{logged}{logical. Return log-likelihood?}

\item{optim_friendly}{logical. avoid -Inf in prior?}
}
\value{
(log-)prior density
}
\description{
Evaluate flat prior density for gpd parameters
}
\examples{
gpd_flat_prior(nu = 0.1,xi = 0)
}
