% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_pareto_likelihoods.R
\name{gpd_likelihood}
\alias{gpd_likelihood}
\title{Calculate generalised Pareto likelihoods, log-likelihoods and negative log-likelihoods.}
\usage{
gpd_likelihood(
  x,
  sig = NULL,
  nu = NULL,
  xi,
  mu,
  logged = FALSE,
  negative = FALSE,
  combined = TRUE
)
}
\arguments{
\item{x}{vector of observed values}

\item{sig}{vector of GPD scale parameters. Specify exactly one of sig and nu.}

\item{nu}{vector of alternative GPD scale parameters. Specify exactly one of sig and nu.}

\item{xi}{vector of GPD shape parameters.}

\item{mu}{vector of GPD thresholds.}

\item{logged}{Logical, stating wheter to return logged value. Default is FALSE.}

\item{negative}{Logical, stating wheter to return negated value. Default is FALSE.}

\item{combined}{Logical, stating wheter to return combined value. Default is TRUE. Liklihood values are combined by multiplication, log-likelihood values are combined by addition.}
}
\value{
vector of Liklihood, log-likelihood or negative log-likelihood value(s).
}
\description{
Calculate generalised Pareto likelihoods, log-likelihoods and negative log-likelihoods.
}
