% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_gpd_to_gauss.R
\name{gpd_to_gaus}
\alias{gpd_to_gaus}
\title{Conversion from GPD to Gaussian margin in R}
\usage{
gpd_to_gaus(x, sig, xi, mu)
}
\arguments{
\item{x}{Numeric vector of gpd observations}

\item{sig}{gpd alt-scale parameter}

\item{xi}{gpd shape parameter}

\item{mu}{gpd threshold}
}
\value{
NULL prints result of Cpp function to put x_gpd on gaussian margins
}
\description{
Conversion from GPD to Gaussian margin in R
}
