% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_gpd_to_gauss.R
\name{gpd_to_gaus_tail}
\alias{gpd_to_gaus_tail}
\title{Conversion from GPD to Gaussian margin for tail values in R (internal function)}
\usage{
gpd_to_gaus_tail(x, sig, xi, mu)
}
\arguments{
\item{x}{Numeric vector of gpd observations}

\item{sig}{alt-scale parameter}

\item{xi}{gpd shape parameter}

\item{mu}{gpd threshold}
}
\value{
x on gaussian margins
}
\description{
Conversion from GPD to Gaussian margin for tail values in R (internal function)
}
