% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_pareto_distribution.R
\name{llh_gpd}
\alias{llh_gpd}
\title{Calculate the log-likelihood of iid generalised Pareto data}
\usage{
llh_gpd(x, nuxi, mu, neg = FALSE)
}
\arguments{
\item{x}{Data vector}

\item{nuxi}{Parameters of GPD}

\item{mu}{Threshold of GPD}

\item{neg}{Logical. Return negative log-likelihood?}
}
\value{
(negative) log-likelihood at (nu,xi)
}
\description{
Calculate the log-likelihood of iid generalised Pareto data
}
