% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_pareto_distribution.R
\name{maxLikelihoodGPD}
\alias{maxLikelihoodGPD}
\title{Find the maximum likelihood estimate for iid generalised Pareto data}
\usage{
maxLikelihoodGPD(nuxi, x, mu, hessian = FALSE, ...)
}
\arguments{
\item{nuxi}{Starting point for optimisation of the likelihood function}

\item{x}{Observations from the generalised pareto distribution}

\item{mu}{threshold for the generealised pareto distribution}

\item{hessian}{Logical. Return hessian in output? Default is false.}

\item{...}{Additional arguments to optim.}
}
\value{
list of maximum likelihood estimate and log-likelihood value at the mle.
}
\description{
Find the maximum likelihood estimate for iid generalised Pareto data
}
\examples{
dat <- rgpd(n = 300, nu = 1.2, shape = 0, mu = 0)
maxLikelihoodGPD(x = dat, mu = 0, nuxi = c(0.1,0))
}
