% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_conversion.R
\name{nu2pow}
\alias{nu2pow}
\title{Convert parameters from gpd (nu, xi) to pwoer law (c,p)}
\usage{
nu2pow(nu, xi)
}
\arguments{
\item{nu}{gpd proxy scale}

\item{xi}{gpd shape}
}
\value{
vector of converted parameters
}
\description{
Convert parameters from gpd (nu, xi) to pwoer law (c,p)
}
\author{
Zak Varty
}
