% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_conversion.R
\name{nu2sig}
\alias{nu2sig}
\title{Convert parameters from gpd (nu,xi) to gpd (sigma,xi)}
\usage{
nu2sig(nu, xi)
}
\arguments{
\item{nu}{gpd proxy scale}

\item{xi}{gpd shape}
}
\value{
vector of converted parameters
}
\description{
Convert parameters from gpd (nu,xi) to gpd (sigma,xi)
}
\author{
Zak Varty
}
