% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_conversion.R
\name{Omori2cGPD}
\alias{Omori2cGPD}
\title{Convert ETAS parameter vector from Omori-law to cGPD parameterisation.}
\usage{
Omori2cGPD(OmoriPars, Mbar, M0)
}
\arguments{
\item{OmoriPars}{Vector of Omori-law ETAS parameters (mu, K, alpha, c, p)}

\item{Mbar}{Mean magnitude of catalogue}

\item{M0}{Threshold magnitude of catalogue}
}
\value{
Vector of cGPD parameters (mu, C, alpha, nu_t, xi_t)
}
\description{
Convert ETAS parameter vector from Omori-law to cGPD parameterisation.
}
