% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_pareto_distribution.R
\name{pgpd}
\alias{pgpd}
\title{Generalised Pareto Distribution}
\usage{
pgpd(q, shape, scale = NULL, nu = NULL, mu = 0)
}
\arguments{
\item{q}{vector of quantiles.}

\item{shape}{shape parameter (xi)}

\item{scale}{scale parameter (sigma)}

\item{nu}{alternative scale parameter: nu = sigma/(1+xi)}

\item{mu}{location parameter}
}
\value{
Probability of the GPD X<=q
}
\description{
Cumulative density function of the GPD specified in terms of (sigma,xi) or (nu,xi).
Improvement over evir function as it returns an error if the (implied) shape
parameter is non-positive. Also properly handles cases where and xi=0 or p<mu.
}
\examples{
pgpd(q = c(-1,1.5,3), shape = 1, scale = 1)
pgpd(q = 1.5, shape = c(0,-1), scale = c(0.1,1))
}
\author{
Zak Varty
}
