% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_conversion.R
\name{pow2nu}
\alias{pow2nu}
\title{Convert Paramters from powerlaw (c,p) to gpd (nu,xi)}
\usage{
pow2nu(c, p)
}
\arguments{
\item{c}{power law length scale}

\item{p}{power law decay}
}
\value{
vector of converted parameters
}
\description{
Convert Paramters from powerlaw (c,p) to gpd (nu,xi)
}
\author{
Zak Varty
}
