% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_conversion.R
\name{pow2sig}
\alias{pow2sig}
\title{Convert prameters from powerlaw (c,p) to gpd (sigma,xi)}
\usage{
pow2sig(c, p)
}
\arguments{
\item{c}{power law length scale}

\item{p}{power law decay}
}
\value{
vector of converted parameters
}
\description{
Convert prameters from powerlaw (c,p) to gpd (sigma,xi)
}
\author{
Zak Varty
}
