% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_pareto_distribution.R
\name{qgpd}
\alias{qgpd}
\title{Generalised Pareto Distribution}
\usage{
qgpd(p, shape, scale = NULL, nu = NULL, mu = 0)
}
\arguments{
\item{p}{vector of quantiles.}

\item{shape}{shape parameter (xi)}

\item{scale}{scale parameter (sigma)}

\item{nu}{alternative scale parameter: nu = sigma/(1+xi)}

\item{mu}{location parameter}
}
\value{
Probability of the GPD X<=x
}
\description{
Cumulative density function of the GPD specified in terms of (sigma,xi) or (nu,xi).
Improvement over evir function as it returns an error if the (implied) shape
parameter is non-positive. Also properly handles cases where and xi=0 or p is not a valid
probability.
}
\examples{
qgpd(p = 0.5, shape = 0.5, scale = 0.5)
\dontrun{ qgpd(p = -0.1, shape = 0, scale = 1, mu = 0.1) }
}
\author{
Zak Varty
}
