% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrETAS_helpers.R
\name{rbranch}
\alias{rbranch}
\title{Sample uniformly from the set of branching vectors of length n.}
\usage{
rbranch(n)
}
\arguments{
\item{n}{integer. Number of elements in branching vector.}
}
\value{
Vector of length n that is a valid branching vector.
}
\description{
Sample uniformly from the set of branching vectors of length n.
}
\examples{
rbranch(n = 100)
}
