% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_pareto_distribution.R
\name{rgpd}
\alias{rgpd}
\title{Generalised Pareto Distribution}
\usage{
rgpd(n, shape, scale = NULL, nu = NULL, mu = 0)
}
\arguments{
\item{n}{sample size.}

\item{shape}{shape parameter (xi).}

\item{scale}{scale parameter (sigma).}

\item{nu}{alternative scale parameter.}

\item{mu}{location parameter.}
}
\value{
Random sample from generalised pareto distirbution.
}
\description{
Sample the GPD specified in terms of (sigma,xi) or (nu,xi).
Improvement over evir function as it returns an error if the (implied) shape
parameter is non-positive. Also properly handles cases where and xi=0.
}
\examples{
rgpd(n = 100, shape = 0, scale = 1:100)
}
\author{
Zak Varty
}
