% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_conversion.R
\name{sig2nu}
\alias{sig2nu}
\title{Convert prameters from gpd (sigma,xi) to gpd (nu,xi)}
\usage{
sig2nu(sig, xi)
}
\arguments{
\item{sig}{gpd scale}

\item{xi}{gpd shape}
}
\value{
vector of converted parameters
}
\description{
Convert prameters from gpd (sigma,xi) to gpd (nu,xi)
}
\author{
Zak Varty
}
