% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_conversion.R
\name{sig2pow}
\alias{sig2pow}
\title{Convert prameters from gpd (sigma,xi) to powerlaw (c,p)}
\usage{
sig2pow(sig, xi)
}
\arguments{
\item{sig}{gpd scale}

\item{xi}{gpd shape}
}
\value{
vector of converted parameters
}
\description{
Convert prameters from gpd (sigma,xi) to powerlaw (c,p)
}
\author{
Zak Varty
}
