% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateETAScorr.R
\name{simulateETAScorr}
\alias{simulateETAScorr}
\title{Simulates synthetic data from the ETAS model}
\usage{
simulateETAScorr(
  mu,
  K,
  alpha,
  gpd_t,
  rho = 0,
  gpd_m0,
  gpd_m1 = NULL,
  m_0,
  t_max,
  displayOutput = TRUE
)
}
\arguments{
\item{mu}{Parameter of the ETAS model as described above.}

\item{K}{Parameter of the ETAS model as described above.}

\item{alpha}{Parameter of the ETAS model as described above.}

\item{gpd_t}{Parameter vector (nu_t,xi_t) for generalised Pareto distribution of aftershock delays}

\item{rho}{Correlation coefficient linking background and triggered magnitudes.}

\item{gpd_m0}{Parameter vector (nu_m0, xi_m0) for generalised pareto distribution of background magnitudes.}

\item{gpd_m1}{Parameter vector (nu_m1, xi_m1) for generalised pareto distribution of triggered magnitudes.}

\item{m_0}{Magnitude of completeness.}

\item{t_max}{Length of the time window [0,t_max] to simulate the catalog over.}

\item{displayOutput}{If TRUE then prints the number of earthquakes simulated so far.}
}
\value{
A list consisting of
  \item{ts}{The simulated earthquake times}
  \item{ms}{The simulated earthquake magnitudes}
  \item{b}{The simulated branching structure, where branching[i] is the index of the earthquake that triggered earthquake i, or 0 if earthquake i is a background event}
}
\description{
This function simulates sample data from the ETAS model over a particular interval [0,T].
The Epidemic Type Aftershock Sequence (ETAS) model is widely used to quantify the degree of seismic activity in a geographical region, and to forecast the occurrence of future mainshocks and aftershocks (Ross 2016). The temporal ETAS model is a point process where the probability of an earthquake occurring at time \eqn{t} depends on the previous seismicity \eqn{H_t}{Ht}, and is defined by the conditional intensity function:
\deqn{ \lambda(t|H_t) = \mu + \sum_{t[i] < t}  \kappa(m[i]|K,\alpha) h(t[i]|c,p)}{ \lambda(t|Ht) = \mu + \sum \kappa(m[i]|K,\alpha) h(t[i]|c,p)}
where
\deqn{\kappa(m_i|K,\alpha) = Ke^{\alpha \left( m_i-M_0 \right)}}{\kappa(m[i]|K,\alpha) = K * exp(\alpha(m[i]-M0))}
and
\deqn{ h(t_i|c,p) = \frac{(p-1)c^{p-1}}{(t-t_i+c)^p}}{ h(t[i]|c,p) = (p-1) * c^(p-1) * (t-t[i]+c)^(-p)}
where the summation is over all previous earthquakes that occurred in the region, with the i'th such earthquake occurring at time \eqn{ t_i}{t[i]} and having magnitude \eqn{ m_i}{m[i]}. The quantity \eqn{M_0}{M0} denotes the magnitude of completeness of the catalog, so that \eqn{m_i \geq  M_0}{m[i] \ge M0} for all i. The temporal ETAS model has 5 parameters: \eqn{\mu} controls the background rate of seismicity, \eqn{ K} and \eqn{ \alpha} determine the productivity (average number of aftershocks) of an earthquake with magnitude \eqn{m}, and \eqn{c} and \eqn{ p}  are the parameters of the Modified Omori Law (which has here been normalized to integrate to 1) and represent the speed at which the aftershock rate decays over time. Each mainshock is assumed to have a magnitude which is an independent draw from a Generalised Pareto Distribution, Aftershocks can follow a separate Generalised Pareto Distribution and can be made dependent on the magnitude of the triggering event through the parameter rho.
\deqn{}
This function simulates sample data from the ETAS model over a particular interval [0,T]. Background and aftershock magnitudes follow marginal generalised pareto distributions linked by a gaussian copula.
}
\author{
Zak Varty
}
